+++ lib/money/ddl.ex
@@ -77,6 +77,38 @@ defmodule Money.DDL do
   end
 
   @doc """
+  Returns the SQL string which when executed will
+  define a `+` operator for the `money_with_currency`
+  data type.
+
+  ## Arguments
+
+  * `db_type`: the type of the database for which the SQL
+    string should be returned.  Defaults to `:postgres` which
+    is currently the only supported database type.
+
+  """
+  def define_plus_operator(db_type \\ @default_db) do
+    read_sql_file(db_type, "define_plus_operator.sql")
+  end
+
+  @doc """
+  Returns the SQL string which when executed will
+  drop the `+` operator for the `money_with_currency`
+  data type.
+
+  ## Arguments
+
+  * `db_type`: the type of the database for which the SQL
+    string should be returned.  Defaults to `:postgres` which
+    is currently the only supported database type.
+
+  """
+  def drop_plus_operator(db_type \\ @default_db) do
+    read_sql_file(db_type, "drop_plus_operator.sql")
+  end
+
+  @doc """
   Returns a string that will Ecto `execute` each SQL
   command.