+++ lib/money_sql.ex
@@ -1,4 +1,5 @@
 defmodule Money.Sql do
   @moduledoc false
-  
+  #require Money
+
 end