+++ priv/SQL/postgres/drop_plus_operator.sql
@@ -1,0 +1,4 @@
+DROP OPERATOR + (money_with_currency, money_with_currency);
+
+
+DROP FUNCTION IF EXISTS money_add(money_1 money_with_currency, money_2 money_with_currency);